requirements.txt
python-telegram-bot[ext]
pyyaml
groq
duckduckgo-search
telegraph[aio]
markdown-it-py
python-dotenv